Restructured and added automation with frames to make everything visible in one place

+++ demo/0-reset_databases/index.php
@@ -1,18 +1,24 @@
 <?php
 
-error_reporting(E_ALL);
-ini_set('display_errors',1);
+if( isset($_POST['submit']) ) {
 
-$db = new PDO('mysql:host=localhost;dbname=understanding', 'root', 'root');
+	error_reporting(E_ALL);
+	ini_set('display_errors',1);
 
-$db->exec("DROP  TABLE user");
-$db->exec("CREATE TABLE user (
-	id integer primary key auto_increment,
-	username text,
-	password text,
-	admin integer
-)");
+	$db = new PDO('mysql:host=localhost;dbname=understanding', 'root', 'root');
 
-$db->exec("INSERT INTO user (username, password, admin) VALUES ('ben', 'password', 0)");
-$db->exec("INSERT INTO user (username, password, admin) VALUES ('administrator', 'complex', 1)");
+	$db->exec("DROP  TABLE user");
+	$db->exec("CREATE TABLE user (
+		id integer primary key auto_increment,
+		username text,
+		password text,
+		admin integer
+	)");
 
+	$db->exec("INSERT INTO user (username, password, admin) VALUES ('ben', 'password', 0)");
+	$db->exec("INSERT INTO user (username, password, admin) VALUES ('administrator', 'complex', 1)");
+
+	echo "DB Reset<br/><br/>";
+}
+?>
+<form action="#" method="post"><input type="submit" name="submit" value="Reset Databases"/></form>
